requirements.txt
aiogram
feedparser
httpx
lxml
python-dotenv
requests
beautifulsoup4
newspaper3k 